pyproject.toml
[project]
name = "langgraph-tool-experiments"
version = "0.1.0"
description = ""
authors = [
    {name = "Paul Williamson",email = "paulw.900430@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "langgraph-cli[inmem] (>=0.2.10,<0.3.0)",
    "langgraph-sdk (>=0.1.69,<0.2.0)",
    "langchain-openai (>=0.3.17,<0.4.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "mypy (>=1.15.0,<2.0.0)",
    "langchain-mcp-adapters (>=0.1.0,<0.2.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "langmem (>=0.0.25,<0.0.26)"
]

[tool.poetry]
packages = [
    { include = "src/agent" },
    { include = "src/tools" },
    { include = "src/memory" },
    { include = "src/embeddings" },
    { include = "src/mcp_servers" }
]

[tool.isort]
multi_line_output = 3
force_grid_wrap = 0
include_trailing_comma = true
use_parentheses = true
line_length = 88

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]  # Enable common error & formatting rules
ignore = ["E501"]  # Ignore line length violations

[tool.black]
line_length = 88
target_version = ['py38', 'py39', 'py310']


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
